requirements.txt
pyyaml
paho-mqtt
python-kasa
janus
